requirements.txt
tensorflow==2.20.0
tf-keras
numpy
gradio
google-generativeai
